Fix location of dependency "architecture-rules"
The test-time dependency "architecture-rules" used to be located in a
Maven repository that was self-hosted on Google Code. Since Google Code
is now defunct, the dependency can no longer be resolved.

The library "architecture-rules" is still available from a repository
run by Smartics. Use this repository instead. Since it provides a more
recent version (3.0.0-rc1 instead of 2.1.1), minor changes to the unit
test code are necessary.